requirements.txt
Flask==2.1.0
Flask-SQLAlchemy==3.0.2
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.1
Flask-Limiter==3.1.0
SQLAlchemy==1.4.46
pandas==1.5.3
python-dotenv==1.0.0
requests==2.28.2
spotipy==2.23.0
gunicorn==20.1.0
psycopg2-binary==2.9.5